backend/requirements.txt
awswrangler
boto3
cryptography
fastapi
gunicorn
httpx
jinja2
pandas
psutil
pyarrow
pyjwt
pydantic[email]
python-dotenv
python-multipart
uvicorn
requests
strenum
pytest-order

